// src/app/models/gas.ts
export class Gas {
  id: number;
  acquired: number;
  fast: number;
  fastest: number;
  safeLow: number;
  average: number;
  // tslint:disable-next-line:variable-name
  block_time: number;
  blockNum: number;
  speed: number;
  safeLowWait: number;
  avgWait: number;
  fastWait: number;
  fastestWait: number;


}